!B (CE-18437) (Renderer, Streaming) Fixed tiling mode size calculation !B (Renderer) Fixed wrong assert when adding lights
Copied from Perforce
 Change: 1886890

+++ Code/CryEngine/RenderDll/Common/Textures/ITextureStreamer.cpp
@@ -99,9 +99,10 @@ size_t ITextureStreamer::StatsComputeRequiredMipMemUsage()
 
 		bool bStale = StatsWouldUnload(tp);
 		{
-			int nPersMip = tp->m_nMips - tp->m_CacheFileHeader.m_nMipsPersistent;
-			int nReqMip = tp->m_bForceStreamHighRes ? 0 : (bStale ? nPersMip : tp->GetRequiredMip());
-			int nMips = tp->GetNumMips();
+			int8 nPersMip = tp->m_nMips - tp->m_CacheFileHeader.m_nMipsPersistent;
+			int8 nReqMip = tp->m_bForceStreamHighRes ? 0 : (bStale ? nPersMip : tp->GetRequiredMip());
+			int8 nMips = tp->GetNumMips();
+
 			nReqMip = min(nReqMip, nPersMip);
 
 			uint32 nWantedSize = tp->StreamComputeSysDataSize(nReqMip);
