+++ Code/CryEngine/RenderDll/Common/Textures/PlanningTextureStreamer.h
@@ -67,10 +67,10 @@ struct SPlanningSortState
 	size_t               nStreamLimit;
 	int32                arrRoundIds[MAX_PREDICTION_ZONES];
 	int                  nFrameId;
-	int                  nBias;
-	int                  fpMinBias;
-	int                  fpMaxBias;
-	int                  fpMinMip;
+	int16                nBias;
+	int16                fpMinBias;
+	int16                fpMaxBias;
+	int16                fpMinMip;
 	SPlanningMemoryState memState;
 
 	// In/Out
@@ -91,7 +91,7 @@ struct SPlanningScheduleState
 {
 	int                    nFrameId;
 
-	int                    nBias;
+	int16                  nBias;
 	SPlanningMemoryState   memState;
 
 	TPlanningTextureReqVec requestList;
@@ -150,7 +150,7 @@ struct SPlanningTextureOrderKey
 	bool   IsVisible() const            { return (nKey & (1 << 29)) == 0; }
 	bool   IsInZone(int z) const        { return (nKey & (1 << (28 - z))) == 0; }
 	bool   IsPrecached() const          { return (nKey & ((1 << 31) | (1 << 28) | (1 << 27))) != ((1 << 31) | (1 << 28) | (1 << 27)); }
-	int    GetFpMinMipCur() const       { return static_cast<int16>(static_cast<int>(nKey & 0xffff) - PackedFpBias); }
+	int16  GetFpMinMipCur() const       { return static_cast<int16>(static_cast<int>(nKey & 0xffff) - PackedFpBias); }
 	uint16 GetFpMinMipCurBiased() const { return (uint16)nKey; }
 
 	SPlanningTextureOrderKey() {}
@@ -188,7 +188,7 @@ struct SPlanningTextureOrderKey
 struct SPlanningRequestIdent
 {
 	SPlanningRequestIdent() {}
-	SPlanningRequestIdent(uint32 nSortKey, int nKey, int nMip)
+	SPlanningRequestIdent(uint32 nSortKey, int nKey, int8 nMip)
 		: nSortKey(nSortKey)
 		, nKey(nKey)
 		, nMip(nMip)
@@ -246,8 +246,8 @@ private:
 
 	void                 Job_UpdateMip(CTexture* pTexture, const float fMipFactor, const int nFlags, const int nUpdateId);
 	void                 Job_Sort();
-	int                  Job_Bias(SPlanningSortState& sortState, SPlanningTextureOrderKey* pKeys, size_t nNumPrecachedTexs, size_t nStreamLimit);
-	size_t               Job_Plan(SPlanningSortState& sortState, const SPlanningTextureOrderKey* pKeys, size_t nTextures, size_t nNumPrecachedTexs, size_t nBalancePoint, int nMinMip, int fpSortStateBias);
+	int16                Job_Bias(SPlanningSortState& sortState, SPlanningTextureOrderKey* pKeys, size_t nNumPrecachedTexs, size_t nStreamLimit);
+	size_t               Job_Plan(SPlanningSortState& sortState, const SPlanningTextureOrderKey* pKeys, size_t nTextures, size_t nNumPrecachedTexs, size_t nBalancePoint, int8 nMinMip, int16 fpSortStateBias);
 	void                 Job_InitKeys(SPlanningTextureOrderKey* pKeys, CTexture** pTexs, size_t nTextures, int nFrameId, const int nZoneIds[]);
 	void                 Job_CommitKeys(CTexture** pTextures, SPlanningTextureOrderKey* pKeys, size_t nTextures);
 	void                 Job_ConfigureSchedule();
@@ -255,14 +255,14 @@ private:
 
 private:
 
-	bool TryBegin_FromDisk(CTexture* pTex, uint32 nTexPersMip, uint32 nTexWantedMip, uint32 nTexAvailMip, int nBias, int nBalancePoint,
+	bool TryBegin_FromDisk(CTexture* pTex, uint32 nTexPersMip, uint32 nTexWantedMip, uint32 nTexAvailMip, int16 nBias, int nBalancePoint,
 	                       TStreamerTextureVec& textures, TStreamerTextureVec& trimmable, ptrdiff_t& nMemFreeLower, ptrdiff_t& nMemFreeUpper, int& nKickIdx,
 	                       int& nNumSubmittedLoad, size_t& nAmtSubmittedLoad);
 
 #if defined(TEXSTRM_TEXTURECENTRIC_MEMORY)
 	bool      TrimTexture(int nBias, TStreamerTextureVec& trimmable, STexPool* pPrioritise);
 #endif
-	ptrdiff_t TrimTextures(ptrdiff_t nRequired, int nBias, TStreamerTextureVec& trimmable);
+	ptrdiff_t TrimTextures(ptrdiff_t nRequired, int16 nBias, TStreamerTextureVec& trimmable);
 	ptrdiff_t KickTextures(CTexture** pTextures, ptrdiff_t nRequired, int nBalancePoint, int& nKickIdx);
 	void      Job_CheckEnqueueForStreaming(CTexture* pTexture, const float fMipFactor, bool bHighPriority);
 
@@ -284,7 +284,7 @@ private:
 
 	SPlanningScheduleState m_schedule;
 
-	int                    m_nBias;
+	int16                  m_nBias;
 	int                    m_nStreamAllocFails;
 	bool                   m_bOverBudget;
 	size_t                 m_nPrevListSize;
