+++ Code/CryEngine/RenderDll/XRenderD3D9/D3DTextureStreamPool.cpp
@@ -111,7 +111,7 @@ STexPool* CTextureStreamPoolMgr::GetPool(const STextureLayout& pLayout)
 	if (pLayout.m_bIsSRGB)
 		d3dFmt = DeviceFormats::ConvertToSRGB(d3dFmt);
 
-	TexturePoolKey poolKey((uint16)pLayout.m_nWidth, (uint16)pLayout.m_nHeight, d3dFmt, (uint8)pLayout.m_eTT, (uint8)pLayout.m_nMips, (uint16)pLayout.m_nArraySize);
+	TexturePoolKey poolKey(pLayout.m_nWidth, pLayout.m_nHeight, d3dFmt, pLayout.m_eTT, pLayout.m_nMips, pLayout.m_nArraySize);
 
 	TexturePoolMap::iterator it = m_TexturesPools.find(poolKey);
 	if (it != m_TexturesPools.end())
@@ -251,12 +251,12 @@ void CTextureStreamPoolMgr::ReleaseItem(STexPoolItem* pItem)
 #endif
 }
 
-STexPool* CTextureStreamPoolMgr::GetOrCreatePool(int nWidth, int nHeight, int nMips, int nArraySize, D3DFormat eTF, ETEX_Type eTT)
+STexPool* CTextureStreamPoolMgr::GetOrCreatePool(uint16 nWidth, uint16 nHeight, int8 nMips, uint16 nArraySize, D3DFormat eTF, ETEX_Type eTT)
 {
 	STexPool* pPool = NULL;
 
 	assert((eTT != eTT_Cube && eTT != eTT_CubeArray) || !(nArraySize % 6));
-	TexturePoolKey poolKey((uint16)nWidth, (uint16)nHeight, eTF, (uint8)eTT, (uint8)nMips, (uint16)nArraySize);
+	TexturePoolKey poolKey(nWidth, nHeight, eTF, eTT, nMips, nArraySize);
 
 	TexturePoolMap::iterator it = m_TexturesPools.find(poolKey);
 	if (it == m_TexturesPools.end())
@@ -389,7 +389,7 @@ void CTextureStreamPoolMgr::GarbageCollect(size_t* nCurTexPoolSize, size_t nLowe
 			ps.nWidth = pPool->m_Width;
 			ps.nHeight = pPool->m_Height;
 			ps.nMips = pPool->m_nMips;
-			ps.nFormat = pPool->m_eFormat;
+			ps.nFormat = uint8(pPool->m_eFormat);
 			ps.eTT = pPool->m_eTT;
 
 			ps.nInUse = pPool->m_nItems - pPool->m_nItemsFree;
