+++ Code/CryEngine/RenderDll/XRenderD3D9/DeviceManager/DeviceResources.h
@@ -472,7 +472,7 @@ public:
 
 	static uint32 TextureDataSize(D3DBaseView* pView);
 	static uint32 TextureDataSize(D3DBaseView* pView, const uint numRects, const RECT* pRects);
-	static uint32 TextureDataSize(uint32 nWidth, uint32 nHeight, uint32 nDepth, uint32 nMips, uint32 nSlices, const ETEX_Format eTF, ETEX_TileMode eTM, uint32 eFlags);
+	static uint32 TextureDataSize(uint32 nWidth, uint32 nHeight, uint32 nDepth, int8 nMips, uint32 nSlices, const ETEX_Format eTF, ETEX_TileMode eTM, uint32 eFlags);
 
 #if DEVRES_USE_PINNING
 	void* WeakPin();
